feat: Added basic dependency injection for ExpansionProvider
This allows certain expansions to obtain data from the platform on which they are running where there is no static context for their platform or plugin/mod instances.
This feature will be used by a minority of expansions, but it eliminates the need to create expansions that load as plugins, as in MiniPlaceholders v1 and v2.