MiniPlaceholders V3: Placeholders Refactor (#256)
* refactor: Changed the way of processing and creating placeholders, now based on the MiniMessage parsing context

* fix: Fixed version string

* feat: Added ability to obtain placeholders directly

resolves #174

* feat: Implement GlobalPlaceholder instance

* feat: Reimplemented support for ForwardingAudiences

* test: Fixed resolver equality tests

* test: Implemented some test based on recent API changes

* docs: Added some documentation

* docs: Added some placeholders documentation

* test: Added complex tag tests

* fix: Fixed javadoc generation

* feat: Implement Expansion Providers

resolves #9

* fix: Apply changes from #265

* feat: Removed default expansions

This allows more control by server admins to be able to control which Expansions are used on a server. The default expansions will be included in external expansions.

* docs: A few advances in documentation

This commit took over three months to publish...

* feat: Added error messages to expansion providers loading process

* build: Updated to Gradle 9.0.0

* docs: Finished javadocs documentation

build: Use gradle configuration cache

* fix: Fixed dynamic expansions loading

* feat: Added basic dependency injection for ExpansionProvider

This allows certain expansions to obtain data from the platform on which they are running where there is no static context for their platform or plugin/mod instances.
This feature will be used by a minority of expansions, but it eliminates the need to create expansions that load as plugins, as in MiniPlaceholders v1 and v2.

* fix: Expansions initialization on Paper